feat - #101 앱관련 도메인 업데이트
